LDpair is now a close-to drop-in replacement for the orig
The only API difference is the way the the request parameter is handled,
in the new version, a full file path is required, in the old API, only a
string was required. Also, the file paths for the 1000genomes and dbSNP
paths are changed to match my environment.

The non-API changes include python 2/3 dual compatability, a drastic
reduction in the amount of duplicate code, and increased documentation,
plus the option to return a simple dictionary.

These changes allow me to use my own code to call this function, without
taking anything away from the original functionality.